pyproject.toml
[project]
name = "workflow-orchestrator"
version = "1.0.0"
description = "AI agent that orchestrates end-to-end data pipeline workflows with retry logic, fallback chains, premium HTML reporting, and structured execution logging."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "openai>=1.57.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"